TensorFlow_implementation/requirements.txt
Pillow
tqdm
matplotlib
tensorflow